Split load_pipeline.py into stages/ package + concurrency/gate.py
Extract the six post-import concerns from the 1592-line load_pipeline.py
blob into a vtscore/datasets/stages/ package (clipper, embedding, finalize,
projection, registry, plus a _common leaf for shared constants and
_origin_to_str), and move the generic ConcurrencyGate primitive to
vtscore/concurrency/gate.py. load_pipeline.py keeps the background-thread
orchestration and request-field parsing helpers. Dependency DAG is one-way
(orchestrator -> stages -> _common), no import cycle.

Update all importers (routes, tests, monkeypatch targets) to the new
canonical module homes, plus docstring cross-references and the
architecture / plan docs.
