+++ tests/detectors/test_clipper_workflow.py
@@ -114,7 +114,7 @@ class TestApplyClipperThumbnails:
     find/label list shows the clip's range, not the parent media."""
 
     def test_audio_clips_get_fresh_waveform_thumbnails(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         parent = _make_audio_media(1, duration=5.1)
         # Pretend the loader generated a thumbnail from the full waveform.
@@ -132,7 +132,7 @@ class TestApplyClipperThumbnails:
             assert clip["thumbnail_bytes"][:8] == b"\x89PNG\r\n\x1a\n"
 
     def test_passthrough_clipper_keeps_parent_thumbnail(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         parent = _make_audio_media(1, duration=2.0)
         parent["thumbnail_bytes"] = b"PARENT_WAVEFORM_PNG"
@@ -156,7 +156,7 @@ class TestApplyClipperMD5:
     """Clips must get unique MD5s so collapse_duplicates doesn't merge them."""
 
     def test_audio_clips_get_recomputed_md5s(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips_dict = {1: _make_audio_media(1)}
         parent_md5 = clips_dict[1]["md5"]
@@ -169,7 +169,7 @@ class TestApplyClipperMD5:
             assert clip["md5"] == hashlib.md5(clip["media_bytes"]).hexdigest()
 
     def test_image_clips_get_recomputed_md5s(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips_dict = {1: _make_image_media(1, width=300, height=100)}
         parent_md5 = clips_dict[1]["md5"]
@@ -181,7 +181,7 @@ class TestApplyClipperMD5:
             assert clip["md5"] == hashlib.md5(clip["media_bytes"]).hexdigest()
 
     def test_text_clips_get_unique_md5s(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips_dict = {1: _make_text_media(1)}
         parent_md5 = clips_dict[1]["md5"]
@@ -194,7 +194,7 @@ class TestApplyClipperMD5:
             assert md5 != parent_md5
 
     def test_video_clips_get_unique_md5s(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips_dict = {1: _make_video_media(1, duration=10.0)}
         parent_md5 = clips_dict[1]["md5"]
@@ -208,7 +208,7 @@ class TestApplyClipperMD5:
 
     def test_dedup_preserves_all_clips(self):
         """collapse_duplicates should NOT merge clips from the same parent."""
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
         from vtsearch.state import collapse_duplicates
 
         clips_dict = {1: _make_audio_media(1, duration=5.1)}
@@ -225,7 +225,7 @@ class TestApplyClipperMD5:
         After clipping, each sub-item must get its own MD5, not the
         importer's value for the parent.
         """
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         media = _make_audio_media(1, duration=5.0)
         # Simulate an importer that provides its own MD5
@@ -245,7 +245,7 @@ class TestApplyClipperMD5:
         After clipping, each sub-item must get its own embedding based
         on the clipped content, not the parent embedding.
         """
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         media = _make_audio_media(1, duration=5.0)
         # Tag the parent embedding so we can detect it later
@@ -267,7 +267,7 @@ class TestApplyClipperMD5:
     def test_importer_md5_kept_for_passthrough_clipper(self):
         """Default (pass-through) clippers should NOT replace the
         importer-provided MD5 since no clipping occurred."""
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         media = _make_audio_media(1, duration=5.0)
         media["md5"] = "importer_provided_md5"
@@ -286,7 +286,7 @@ class TestApplyClipperOriginBoundaries:
     """Clip boundaries must be stored in origin params."""
 
     def test_audio_clip_origin_has_boundaries(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips_dict = {1: _make_audio_media(1, duration=5.0)}
         _apply_clipper(clips_dict, "sound_tiling", {"duration": 2.0})
@@ -302,7 +302,7 @@ class TestApplyClipperOriginBoundaries:
             assert float(params["clip_end"]) > float(params["clip_start"])
 
     def test_image_clip_origin_has_clip_box(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips_dict = {1: _make_image_media(1, width=300, height=100)}
         _apply_clipper(clips_dict, "image_tiling")
@@ -316,7 +316,7 @@ class TestApplyClipperOriginBoundaries:
             assert len(box_parts) == 4
 
     def test_video_clip_origin_has_boundaries(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips_dict = {1: _make_video_media(1, duration=10.0)}
         _apply_clipper(clips_dict, "video_tiling", {"duration": 2.0})
@@ -328,7 +328,7 @@ class TestApplyClipperOriginBoundaries:
             assert "clip_end" in params
 
     def test_text_clip_origin_has_clipper_and_index(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips_dict = {1: _make_text_media(1)}
         _apply_clipper(clips_dict, "text_sentence")
@@ -341,7 +341,7 @@ class TestApplyClipperOriginBoundaries:
     def test_audio_clip_origin_stores_clipper_params(self):
         """Clipper parameter values (duration, min_overlap) must be stored
         in origin so cross-dataset resolution can reconstruct the clipper."""
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips_dict = {1: _make_audio_media(1, duration=5.0)}
         _apply_clipper(clips_dict, "sound_tiling", {"duration": 2.5, "min_overlap": 0.5})
@@ -353,7 +353,7 @@ class TestApplyClipperOriginBoundaries:
 
     def test_video_scene_origin_stores_clipper_params(self):
         """Scene clipper's threshold and min_scene_duration must be stored."""
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         # VideoSceneClipper won't produce multiple clips without real video,
         # but the params should still be stored on the single passthrough.
@@ -367,7 +367,7 @@ class TestApplyClipperOriginBoundaries:
 
     def test_default_clipper_stores_no_extra_params(self):
         """Default clippers have no parameter values to store."""
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips_dict = {1: _make_audio_media(1, duration=5.0)}
         _apply_clipper(clips_dict, "sound_default")
@@ -394,7 +394,7 @@ class TestLabelExportWithClips:
         medias.clear()
         try:
             # Set up clipped audio media
-            from vtscore.datasets.load_pipeline import _apply_clipper
+            from vtscore.datasets.stages.clipper import _apply_clipper
 
             clips_dict = {1: _make_audio_media(1, duration=5.0)}
             _apply_clipper(clips_dict, "sound_tiling", {"duration": 2.0})
@@ -437,7 +437,7 @@ class TestLabelImportWithClips:
     """Labels from clipped media can be resolved on the same dataset."""
 
     def test_origin_lookup_matches_clipped_media(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
         from vtscore.state.media_lookup import build_media_lookup, resolve_media_ids
 
         clips_dict = {1: _make_audio_media(1, duration=5.0)}
@@ -456,7 +456,7 @@ class TestLabelImportWithClips:
             assert clip["id"] in matches, f"clip {clip['id']} should be resolvable by origin"
 
     def test_md5_lookup_matches_clipped_media(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
         from vtscore.state.media_lookup import build_media_lookup, resolve_media_ids
 
         clips_dict = {1: _make_audio_media(1, duration=5.0)}
@@ -571,7 +571,7 @@ class TestAPIClipMetadata:
         saved = dict(medias)
         medias.clear()
         try:
-            from vtscore.datasets.load_pipeline import _apply_clipper
+            from vtscore.datasets.stages.clipper import _apply_clipper
 
             clips_dict = {1: _make_audio_media(1, duration=5.0)}
             _apply_clipper(clips_dict, "sound_tiling", {"duration": 2.0})
@@ -595,7 +595,7 @@ class TestAPIClipMetadata:
         saved = dict(medias)
         medias.clear()
         try:
-            from vtscore.datasets.load_pipeline import _apply_clipper
+            from vtscore.datasets.stages.clipper import _apply_clipper
 
             clips_dict = {1: _make_image_media(1, width=300, height=100)}
             _apply_clipper(clips_dict, "image_tiling")
@@ -623,7 +623,7 @@ class TestDefaultClippersNoOp:
     """Default (pass-through) clippers should not break the workflow."""
 
     def test_audio_default_clipper_preserves_media(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         media = _make_audio_media(1, duration=5.0)
         original_md5 = media["md5"]
@@ -634,7 +634,7 @@ class TestDefaultClippersNoOp:
         assert clips_dict[1]["md5"] == original_md5
 
     def test_image_default_clipper_preserves_media(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         media = _make_image_media(1)
         original_md5 = media["md5"]
@@ -645,7 +645,7 @@ class TestDefaultClippersNoOp:
         assert clips_dict[1]["md5"] == original_md5
 
     def test_text_default_clipper_preserves_media(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         media = _make_text_media(1)
         original_md5 = media["md5"]
@@ -656,7 +656,7 @@ class TestDefaultClippersNoOp:
         assert clips_dict[1]["md5"] == original_md5
 
     def test_video_default_clipper_preserves_media(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         media = _make_video_media(1)
         original_md5 = media["md5"]
@@ -676,7 +676,7 @@ class TestMultipleMediasClipped:
     """Clipping multiple medias at once should produce correct results."""
 
     def test_multiple_audio_medias_clipped(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips_dict = {
             1: _make_audio_media(1, duration=5.1),
@@ -691,7 +691,7 @@ class TestMultipleMediasClipped:
             assert clip["md5"] == hashlib.md5(clip["media_bytes"]).hexdigest()
 
     def test_multiple_text_medias_clipped(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips_dict = {
             1: _make_text_media(1, "Foo. Bar. Baz."),
@@ -705,7 +705,7 @@ class TestMultipleMediasClipped:
         assert len(set(md5s)) == 4
 
     def test_clip_ids_are_sequential(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips_dict = {
             1: _make_audio_media(1, duration=5.1),
@@ -726,7 +726,7 @@ class TestApplyClipperProgress:
     """Verify that _apply_clipper reports progress via the on_progress callback."""
 
     def test_progress_reports_clipping_and_embedding_phases(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips_dict = {
             1: _make_audio_media(1, duration=5.1),
@@ -758,7 +758,7 @@ class TestApplyClipperProgress:
 
     def test_no_progress_without_callback(self):
         """Passing on_progress=None doesn't break anything."""
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips_dict = {1: _make_audio_media(1, duration=5.1)}
         _apply_clipper(clips_dict, "sound_tiling", {"duration": 2.0}, on_progress=None)
@@ -828,7 +828,7 @@ class TestClipFieldsInEnrichedExport:
         saved = dict(medias)
         medias.clear()
         try:
-            from vtscore.datasets.load_pipeline import _apply_clipper
+            from vtscore.datasets.stages.clipper import _apply_clipper
 
             clips_dict = {1: _make_audio_media(1, duration=5.0)}
             _apply_clipper(clips_dict, "sound_tiling", {"duration": 2.0})
@@ -861,7 +861,7 @@ class TestClipFieldsInEnrichedExport:
         saved = dict(medias)
         medias.clear()
         try:
-            from vtscore.datasets.load_pipeline import _apply_clipper
+            from vtscore.datasets.stages.clipper import _apply_clipper
 
             clips_dict = {1: _make_image_media(1, width=300, height=100)}
             _apply_clipper(clips_dict, "image_tiling")
