+++ tests/detectors/test_clippers.py
@@ -1884,21 +1884,21 @@ class TestCropFileBytes:
 
 class TestApplyClipper:
     def test_apply_clipper_noop_for_empty_name(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips = {1: {"id": 1, "media_type": "audio", "origin": {"importer": "test", "params": {}}}}
         _apply_clipper(clips, "")
         assert len(clips) == 1
 
     def test_apply_clipper_unknown_name_noop(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         clips = {1: {"id": 1, "media_type": "audio", "origin": {"importer": "test", "params": {}}}}
         _apply_clipper(clips, "nonexistent_clipper")
         assert len(clips) == 1
 
     def test_apply_default_clipper_passthrough(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         media = {"id": 1, "media_type": "audio", "media_bytes": b"fake", "origin": {"importer": "test", "params": {}}}
         clips = {1: media}
@@ -1907,7 +1907,7 @@ class TestApplyClipper:
         assert clips[1]["origin"]["params"]["clipper"] == "sound_default"
 
     def test_apply_clipper_annotates_origin(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         media = {
             "id": 1,
@@ -2177,7 +2177,7 @@ class TestApplyClipperWithParams:
 
     def test_apply_clipper_with_custom_duration(self):
         from vtscore.media.audio.audio_generator import generate_wav
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         # Generate a 10s audio clip
         wav = generate_wav(440, 10.0)
@@ -2195,7 +2195,7 @@ class TestApplyClipperWithParams:
 
     def test_apply_clipper_with_overridden_duration(self):
         from vtscore.media.audio.audio_generator import generate_wav
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         wav = generate_wav(440, 10.0)
         media = {
@@ -2212,7 +2212,7 @@ class TestApplyClipperWithParams:
 
     def test_apply_clipper_params_none_uses_defaults(self):
         from vtscore.media.audio.audio_generator import generate_wav
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         wav = generate_wav(440, 10.0)
         media = {
@@ -2228,7 +2228,7 @@ class TestApplyClipperWithParams:
 
     def test_apply_clipper_with_min_overlap(self):
         from vtscore.media.audio.audio_generator import generate_wav
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         wav = generate_wav(440, 10.0)
         media = {
@@ -2438,7 +2438,7 @@ class TestApplyClipperResolvesAuto:
     clip's origin with the resolved concrete clipper, not the auto one."""
 
     def test_short_dataset_resolves_to_default(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         wav = generate_wav(440, 5.0)
         media = {
@@ -2457,7 +2457,7 @@ class TestApplyClipperResolvesAuto:
         assert clip["origin"]["params"]["clipper"] == "sound_default"
 
     def test_long_dataset_resolves_to_tiling(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         wav = generate_wav(440, 60.0)
         media = {
@@ -2479,7 +2479,7 @@ class TestApplyClipperResolvesAuto:
     def test_mixed_durations_resolve_per_media(self):
         """A short and a long item in the same dataset take different
         branches, and each clip records its own resolved concrete clipper."""
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         short_wav = generate_wav(440, 2.0)
         long_wav = generate_wav(440, 8.0)
@@ -2508,7 +2508,7 @@ class TestApplyClipperResolvesAuto:
         assert resolved_names.count("sound_tiling") == 4
 
     def test_user_threshold_override_propagates(self):
-        from vtscore.datasets.load_pipeline import _apply_clipper
+        from vtscore.datasets.stages.clipper import _apply_clipper
 
         wav = generate_wav(440, 8.0)
         media = {
