+++ vtscore/datasets/ingest.py
@@ -159,7 +159,7 @@ def _boundary_tag(params: dict[str, Any]) -> str:
     the embedding step couldn't produce distinct content bytes) so each
     clip still gets a unique MD5 instead of colliding on the parent's hash.
     Mirrors the encoding in
-    :func:`vtscore.datasets.load_pipeline._fixup_clip_md5_and_embeddings`.
+    :func:`vtscore.datasets.stages.clipper._fixup_clip_md5_and_embeddings`.
     """
     parts: list[str] = []
     for key in ("clipper", "clipper_chain", "clip_start", "clip_end", "clip_box", "clip_index"):
