+++ vtscore/datasets/stages/_common.py
@@ -1,0 +1,40 @@
+"""Shared constants and helpers for the load-pipeline stages.
+
+This is the leaf module of the stages package: it imports nothing from
+:mod:`vtscore.datasets.load_pipeline`, so both the orchestrator and the
+individual stage modules can depend on it without forming an import cycle.
+"""
+
+from __future__ import annotations
+
+# Maps the status strings emitted by inner functions to step numbers.
+# "downloading" covers both download and extraction.
+# "loading" covers model loading and pickle loading.
+# "embedding" covers per-file embedding.
+_STATUS_TO_STEP = {
+    "downloading": 1,
+    "loading": 2,
+    "embedding": 3,
+}
+_TOTAL_LOAD_STEPS = 4  # download, load model, embed, finalize
+
+
+def _origin_to_str(origin: dict | None) -> str:
+    """Convert an origin dict to a human-readable string."""
+    if not origin:
+        return "unknown"
+    importer_name = origin.get("importer", "")
+    if not importer_name:
+        return "unknown"
+
+    from vtscore.datasets.importers import get_importer
+
+    importer = get_importer(importer_name)
+    if importer is not None:
+        return importer.origin_display(origin)
+
+    params = origin.get("params", {})
+    if params:
+        first_val = next(iter(params.values()))
+        return f"{importer_name}:{first_val}"
+    return importer_name
